requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.31
pymysql>=1.1.1
python-dotenv>=1.0.1
plotly>=5.22.0
numpy>=1.26.4
jinja2>=3.1.4
pydantic>=2.8.2
python-multipart>=0.0.9 